+++ system/src/Grav/Common/Data/Blueprint.php
@@ -76,7 +76,9 @@ class Blueprint
         try {
             $this->validateArray($data, $this->nested);
         } catch (\RuntimeException $e) {
-            throw new \RuntimeException(sprintf('<b>Validation failed:</b> %s', $e->getMessage()));
+            $language = self::getGrav()['language'];
+            $message = sprintf($language->translate('FORM.VALIDATION_FAIL', null, true) . ' %s', $e->getMessage());
+            throw new \RuntimeException($message);
         }
     }
 
@@ -449,7 +451,8 @@ class Blueprint
             if (isset($field['validate']['required'])
                 && $field['validate']['required'] === true
                 && empty($data[$name])) {
-                throw new \RuntimeException("Missing required field: {$field['name']}");
+                $value = isset($field['label']) ? $field['label'] : $field['name'];
+                throw new \RuntimeException("Missing required field: {$value}");
             }
         }
     }