+++ system/src/Grav/Common/GPM/Remote/Themes.php
@@ -16,6 +16,8 @@ class Themes extends AbstractPackageCollection
 
     /**
      * Local Themes Constructor
+     * @param bool $refresh
+     * @param callable $callback Either a function or callback in array notation
      */
     public function __construct($refresh = false, $callback = null)
     {