+++ system/src/Grav/Common/Page/Medium/ImageFile.php
@@ -4,10 +4,19 @@ namespace Grav\Common\Page\Medium;
 use Grav\Common\GravTrait;
 use Gregwar\Image\Exceptions\GenerationError;
 use RocketTheme\Toolbox\Event\Event;
+use Gregwar\Image\Image;
 
-class ImageFile extends \Gregwar\Image\Image
+class ImageFile extends Image
 {
     use GravTrait;
+
+    /**
+     * Clear previously applied operations
+     */
+    public function clearOperations()
+    {
+        $this->operations = [];
+    }
 
     /**
      * This is the same as the Gregwar Image class except this one fires a Grav Event on creation of new cached file