+++ system/src/Grav/Common/Plugins.php
@@ -4,7 +4,6 @@ namespace Grav\Common;
 use Grav\Common\Config\Config;
 use Grav\Common\Data\Blueprints;
 use Grav\Common\Data\Data;
-use Grav\Common\GravTrait;
 use Grav\Common\File\CompiledYamlFile;
 use RocketTheme\Toolbox\Event\EventDispatcher;
 use RocketTheme\Toolbox\Event\EventSubscriberInterface;
@@ -122,6 +121,12 @@ class Plugins extends Iterator
 
         // Load default configuration.
         $file = CompiledYamlFile::instance("plugins://{$name}/{$name}.yaml");
+
+        // ensure the plugin exists physically
+        if (!$file->exists()) {
+            return null;
+        }
+
         $obj = new Data($file->content(), $blueprint);
 
         // Override with user configuration.