+++ system/src/Grav/Common/Utils.php
@@ -100,6 +100,9 @@ abstract class Utils
         return (object)array_merge((array)$obj1, (array)$obj2);
     }
 
+    /**
+     * @return array
+     */
     public static function dateFormats()
     {
         $now = new DateTime();
@@ -300,6 +303,19 @@ abstract class Utils
         return $root . implode('/', $ret);
     }
 
+    /**
+     * @param $function
+     *
+     * @return bool
+     */
+    public static function isFunctionDisabled($function)
+    {
+        return in_array($function, explode(',', ini_get('disable_functions')));
+    }
+
+    /**
+     * @return array
+     */
     public static function timezones()
     {
         $timezones = \DateTimeZone::listIdentifiers(\DateTimeZone::ALL);
@@ -327,6 +343,12 @@ abstract class Utils
 
     }
 
+    /**
+     * @param array $source
+     * @param       $fn
+     *
+     * @return array
+     */
     public static function arrayFilterRecursive(Array $source, $fn)
     {
         $result = array();
@@ -346,6 +368,11 @@ abstract class Utils
         return $result;
     }
 
+    /**
+     * @param $string
+     *
+     * @return bool
+     */
     public static function pathPrefixedByLangCode($string)
     {
         $languages_enabled = self::getGrav()['config']->get('system.languages.supported', []);
@@ -357,6 +384,11 @@ abstract class Utils
         return false;
     }
 
+    /**
+     * @param $date
+     *
+     * @return int
+     */
     public static function date2timestamp($date)
     {
         $config = self::getGrav()['config'];
@@ -384,7 +416,111 @@ abstract class Utils
      *
      * @return boolean
      */
-    public static function isPositive($value) {
+    public static function isPositive($value)
+    {
         return in_array($value, [true, 1, '1', 'yes', 'on', 'true'], true);
     }
+
+
+    /**
+     * Generates a nonce string to be hashed. Called by self::getNonce()
+     *
+     * @param string $action
+     * @param bool $plusOneTick if true, generates the token for the next tick (the next 12 hours)
+     *
+     * @return string the nonce string
+     */
+    private static function generateNonceString($action, $plusOneTick = false)
+    {
+        if (isset(self::getGrav()['user'])) {
+            $user = self::getGrav()['user'];
+            $username = $user->username;
+        } else {
+            $username = false;
+        }
+
+        if (!$username) {
+            $username = isset($_SERVER['REMOTE_ADDR']) ? $_SERVER['REMOTE_ADDR'] : '';
+        }
+
+        $token = session_id();
+        $i = self::nonceTick();
+
+        if ($plusOneTick) {
+            $i++;
+        }
+
+        return ( $i . '|' . $action . '|' . $username . '|' . $token );
+    }
+
+    /**
+     * Get the time-dependent variable for nonce creation.
+     *
+     * @todo now a tick lasts a day. Once the day is passed, the nonce is not valid any more. Find a better way
+     *       to ensure nonces issued near the end of the day do not expire in that small amount of time
+     *
+     * @return int the time part of the nonce. Changes once every 24 hours
+     */
+    private static function nonceTick()
+    {
+        $secondsInHalfADay = 60 * 60 * 12;
+        return (int)ceil(time() / ( $secondsInHalfADay ));
+    }
+
+    /**
+     * Get hash of given string
+     *
+     * @param string $data string to hash
+     *
+     * @return string hashed value of $data, cut to 10 characters
+     */
+    private static function hash($data)
+    {
+        $hash = password_hash($data, PASSWORD_DEFAULT);
+        return $hash;
+    }
+
+    /**
+     * Creates a hashed nonce tied to the passed action. Tied to the current user and time. The nonce for a given
+     * action is the same for 12 hours.
+     *
+     * @param string $action the action the nonce is tied to (e.g. save-user-admin or move-page-homepage)
+     * @param bool $plusOneTick if true, generates the token for the next tick (the next 12 hours)
+     *
+     * @return string the nonce
+     */
+    public static function getNonce($action, $plusOneTick = false)
+    {
+        $nonce = self::hash(self::generateNonceString($action, $plusOneTick));
+        $nonce = str_replace('/', 'SLASH', $nonce);
+        return $nonce;
+    }
+
+    /**
+     * Verify the passed nonce for the give action
+     *
+     * @param string $nonce the nonce to verify
+     * @param string $action the action to verify the nonce to
+     *
+     * @return boolean verified or not
+     */
+    public static function verifyNonce($nonce, $action)
+    {
+        $nonce = str_replace('SLASH', '/', $nonce);
+
+        //Nonce generated 0-12 hours ago
+        if (password_verify(self::generateNonceString($action), $nonce)) {
+            return true;
+        }
+
+        //Nonce generated 12-24 hours ago
+        $plusOneTick = true;
+        if (password_verify(self::generateNonceString($action, $plusOneTick), $nonce)) {
+            return true;
+        }
+
+        //Invalid nonce
+        return false;
+    }
+
 }