+++ system/src/Grav/Console/ConsoleCommand.php
@@ -1,0 +1,35 @@
+<?php
+namespace Grav\Console;
+
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Output\OutputInterface;
+
+/**
+ * Class ConsoleCommand
+ *
+ * @package Grav\Console
+ */
+class ConsoleCommand extends Command
+{
+    use ConsoleTrait;
+
+    /**
+     * @param InputInterface  $input
+     * @param OutputInterface $output
+     *
+     * @return int|null|void
+     */
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
+        $this->setupConsole($input, $output);
+        $this->serve();
+    }
+
+    /**
+     *
+     */
+    protected function serve() { }
+
+}